+++ src/main/java/com/emmanuel/domain/models/TaskDetails.java
@@ -2,20 +2,20 @@ package com.emmanuel.domain.models;
 
 public class TaskDetails {
 
-	private final Long userId;
+	private final Integer userId;
 	
 	private final String username;
 
 	private final String userEmail;
 
-	public TaskDetails(Long userId, String username, String userEmail) {
+	public TaskDetails(Integer userId, String username, String userEmail) {
 		super();
 		this.userId = userId;
 		this.username = username;
 		this.userEmail = userEmail;
 	}
 
-	public Long getUserId() {
+	public Integer getUserId() {
 		return userId;
 	}
